+++ apps/mxf2raw/mxf2raw.cpp
@@ -2892,7 +2892,7 @@ int main(int argc, const char** argv)
                 max_samples_per_read = 1920;
 
             // realtime reading
-            uint32_t rt_start = 0;
+            uint64_t rt_start = 0;
             if (realtime)
                 rt_start = get_tick_count();
 
@@ -2900,7 +2900,7 @@ int main(int argc, const char** argv)
             unsigned int gf_retry_count = 0;
             bool gf_read_failure = false;
             int64_t gf_failure_num_read = 0;
-            uint32_t gf_failure_start = 0;
+            uint64_t gf_failure_start = 0;
 
             // read data
             bmx::ByteArray sound_buffer;
